+++ src/vector/vector_trait.rs
@@ -53,9 +53,28 @@ pub trait Vector<S: Scalar>:
     // Associated types.
     // -----------------
 
-    /// Dynamically-sized vector type that is of the same "outer" vector type (i.e. the `Vector`
-    /// part of `Vector<S>` where `S: Scalar`), but where the element type can be any other type
-    /// that implements the [`crate::Scalar`] trait.
+
+    /// Vector type that is of the same "outer" vector type (i.e. the `Vector` part of `Vector<S>`
+    /// where `S: Scalar`), but where the element type can be any other type that implements the
+    /// [`crate::Scalar`] trait.
+    /// 
+    /// # Note
+    /// 
+    /// * This associated type could be either statically or dynamically-sized, as long as it is
+    ///   compatible with this vector type.
+    /// * We recommend that statically-sized vectors choose a compatible statically-sized matrix for
+    ///   this associated type, and the dynamically-sized vectors choose a compatible
+    ///   dynamically-sized matrix for this associated type.
+    /// * For [`ndarray::Array1`], we define this associated type as a [`Vec`]. This is because
+    ///   elements of an [`ndarray::Array1`] must also implement [`ndarray::ScalarOperand`] and
+    ///   [`ndarray::LinalgScalar`], but we cannot apply these trait bounds in the definition of
+    ///   this associated type if we wish to keep [`crate::Vector`] independent of any external
+    ///   crate.
+    type VectorT<T: Scalar>: Vector<T>;
+
+    /// Dynamically-sized vector type that is compatible with this "outer" vector type (i.e. the
+    /// `Vector` part of `Vector<S>` where `S: Scalar`), but where the element type can be any other
+    /// type that implements the [`crate::Scalar`] trait.
     /// 
     /// # Note
     /// 
@@ -63,7 +82,7 @@ pub trait Vector<S: Scalar>:
     /// elements of an [`ndarray::Array1`] must also implement [`ndarray::ScalarOperand`] and
     /// [`ndarray::LinalgScalar`], but we cannot apply these trait bounds in the definition of this
     /// associated type if we wish to keep [`crate::Vector`] independent of any external crate.
-    type GenericVector<T: Scalar>: Vector<T>;
+    type DVectorT<T: Scalar>: Vector<T>;
 
     /// Length-`N` vector type that is of the same "outer" vector type (i.e. the `Vector` part of
     /// `Vector<S>` where `S: Scalar`), but where the elements are of type [`f64`].